Move rewrite stage task to amoro-iceberg module
