+++ amoro-ams/src/main/java/org/apache/amoro/server/optimizing/KeyedTableCommit.java
@@ -30,6 +30,7 @@ import org.apache.amoro.op.OverwriteBaseFiles;
 import org.apache.amoro.op.SnapshotSummary;
 import org.apache.amoro.optimizing.RewriteFilesInput;
 import org.apache.amoro.optimizing.RewriteFilesOutput;
+import org.apache.amoro.optimizing.RewriteStageTask;
 import org.apache.amoro.server.AmoroServiceConstants;
 import org.apache.amoro.table.MixedTable;
 import org.apache.amoro.utils.ContentFiles;
