+++ amoro-ams/src/test/java/org/apache/amoro/server/optimizing/flow/checker/DataConcurrencyChecker.java
@@ -18,7 +18,7 @@
 
 package org.apache.amoro.server.optimizing.flow.checker;
 
-import org.apache.amoro.server.optimizing.RewriteStageTask;
+import org.apache.amoro.optimizing.RewriteStageTask;
 import org.apache.amoro.server.optimizing.UnKeyedTableCommit;
 import org.apache.amoro.server.optimizing.flow.CompleteOptimizingFlow;
 import org.apache.amoro.server.optimizing.flow.DataReader;
