+++ amoro-ams/src/test/java/org/apache/amoro/server/optimizing/flow/checker/FullOptimizingMove2HiveChecker.java
@@ -20,7 +20,7 @@ package org.apache.amoro.server.optimizing.flow.checker;
 
 import org.apache.amoro.optimizing.OptimizingInputProperties;
 import org.apache.amoro.optimizing.OptimizingType;
-import org.apache.amoro.server.optimizing.RewriteStageTask;
+import org.apache.amoro.optimizing.RewriteStageTask;
 import org.apache.amoro.server.optimizing.UnKeyedTableCommit;
 import org.apache.amoro.server.optimizing.flow.view.TableDataView;
 import org.apache.amoro.server.optimizing.plan.OptimizingPlanner;
