+++ amoro-ams/src/test/java/org/apache/amoro/server/optimizing/flow/checker/OptimizingCountChecker.java
@@ -18,7 +18,7 @@
 
 package org.apache.amoro.server.optimizing.flow.checker;
 
-import org.apache.amoro.server.optimizing.RewriteStageTask;
+import org.apache.amoro.optimizing.RewriteStageTask;
 import org.apache.amoro.server.optimizing.UnKeyedTableCommit;
 import org.apache.amoro.server.optimizing.plan.OptimizingPlanner;
 import org.apache.amoro.table.MixedTable;
