+++ amoro-format-iceberg/src/main/java/org/apache/amoro/io/reader/StructLikeFunnel.java
@@ -21,6 +21,7 @@ package org.apache.amoro.io.reader;
 import org.apache.amoro.shade.guava32.com.google.common.hash.Funnel;
 import org.apache.amoro.shade.guava32.com.google.common.hash.PrimitiveSink;
 import org.apache.amoro.utils.SerializationUtil;
+import org.apache.amoro.utils.map.StructLikeWrapperSerializer;
 import org.apache.iceberg.StructLike;
 import org.jetbrains.annotations.NotNull;
 
@@ -31,7 +32,7 @@ public enum StructLikeFunnel implements Funnel<StructLike> {
 
   @Override
   public void funnel(@NotNull StructLike structLike, PrimitiveSink primitiveSink) {
-    StructLike copy = SerializationUtil.StructLikeCopy.copy(structLike);
+    StructLike copy = StructLikeWrapperSerializer.StructLikeCopy.copy(structLike);
     primitiveSink.putBytes(SerializationUtil.kryoSerialize(copy));
   }
 }
